extract listener to a class
