+++ android/beagle/src/main/java/br/com/zup/beagle/android/context/ContextComponentHandler.kt
@@ -22,11 +22,27 @@ import br.com.zup.beagle.core.ServerDrivenComponent
 
 internal class ContextComponentHandler {
 
-    fun handleContext(viewModel: ScreenContextViewModel, view: View, component: ServerDrivenComponent) {
+    fun addContext(viewModel: ScreenContextViewModel, view: View, component: ServerDrivenComponent) {
         if (component is ContextComponent) {
             component.context?.let { context ->
                 viewModel.addContext(view, context)
             }
         }
+    }
+
+    fun addListenerToHandleContext(viewModel: ScreenContextViewModel, view: View) {
+        view.addOnAttachStateChangeListener(object: View.OnAttachStateChangeListener{
+            override fun onViewDetachedFromWindow(v: View?) {
+                v?.let {
+                    viewModel.clearContext(it)
+                }
+            }
+
+            override fun onViewAttachedToWindow(v: View?) {
+                v?.let {
+                    viewModel.linkBindingToContextAndEvaluateThem(it)
+                }
+            }
+        })
     }
 }
