+++ android/beagle/src/main/java/br/com/zup/beagle/android/utils/ViewGroupExtensions.kt
@@ -113,8 +113,6 @@ private fun loadView(
     listener: OnServerStateChanged?
 ) {
     val viewModel = rootView.generateViewModelInstance<GenerateIdViewModel>()
-    val contextViewModel = rootView.generateViewModelInstance<ScreenContextViewModel>()
-
     viewModel.createIfNotExisting(rootView.getParentId())
     val view = viewExtensionsViewFactory.makeBeagleView(rootView).apply {
         serverStateChangedListener = listener
@@ -123,7 +121,6 @@ private fun loadView(
     view.loadCompletedListener = {
         viewGroup.addView(view)
         viewModel.setViewCreated(rootView.getParentId())
-        contextViewModel.linkBindingToContextAndEvaluateThem()
     }
 }
