+++ android/beagle/src/test/java/br/com/zup/beagle/android/context/ContextComponentHandlerTest.kt
@@ -20,12 +20,15 @@ import android.view.View
 import androidx.lifecycle.ViewModelProvider
 import br.com.zup.beagle.android.components.layout.Container
 import br.com.zup.beagle.android.engine.renderer.ActivityRootView
+import br.com.zup.beagle.android.extensions.once
+import br.com.zup.beagle.android.testutil.RandomData
 import br.com.zup.beagle.android.view.viewmodel.ScreenContextViewModel
 import io.mockk.Runs
 import io.mockk.every
 import io.mockk.just
 import io.mockk.mockk
 import io.mockk.mockkConstructor
+import io.mockk.slot
 import io.mockk.unmockkAll
 import io.mockk.verify
 import org.junit.After
@@ -36,7 +39,7 @@ class ContextComponentHandlerTest {
 
     private val rootView = mockk<ActivityRootView>()
     private val view = mockk<View>()
-    private val viewModel = mockk<ScreenContextViewModel>()
+    private val viewModel = mockk<ScreenContextViewModel>(relaxed = true)
 
     private lateinit var contextComponentHandler: ContextComponentHandler
 
@@ -56,17 +59,44 @@ class ContextComponentHandlerTest {
     }
 
     @Test
-    fun handleContext_should_call_addContext_when_component_is_ContextComponent() {
+    fun addContext_should_call_addContext_when_component_is_ContextComponent() {
         // Given
         val component = mockk<Container>()
         val context = mockk<ContextData>()
         every { component.context } returns context
-        every { viewModel.addContext(any(), any()) } just Runs
 
         // When
-        contextComponentHandler.handleContext(viewModel, view, component)
+        contextComponentHandler.addContext(viewModel, view, component)
 
         // Then
         verify(exactly = 1) { viewModel.addContext(view, context) }
+    }
+
+    @Test
+    fun onViewDetachedFromWindow_should_call_clearContext() {
+        // Given
+        val listenerSlot = slot<View.OnAttachStateChangeListener>()
+        every { view.addOnAttachStateChangeListener(capture(listenerSlot)) } just Runs
+
+        // When
+        contextComponentHandler.addListenerToHandleContext(viewModel, view)
+        listenerSlot.captured.onViewDetachedFromWindow(view)
+
+        // Then
+        verify(exactly = once()) { viewModel.clearContext(view) }
+    }
+
+    @Test
+    fun onViewAttachedToWindow_should_call_linkBindingToContextAndEvaluateThem() {
+        // Given
+        val listenerSlot = slot<View.OnAttachStateChangeListener>()
+        every { view.addOnAttachStateChangeListener(capture(listenerSlot)) } just Runs
+
+        // When
+        contextComponentHandler.addListenerToHandleContext(viewModel, view)
+        listenerSlot.captured.onViewAttachedToWindow(view)
+
+        // Then
+        verify(exactly = once()) { viewModel.linkBindingToContextAndEvaluateThem(view) }
     }
 }
